Setup global data store (1/n)
